+++ lightly_studio/src/lightly_studio/models/collection.py
@@ -28,8 +28,11 @@ class CollectionBase(SQLModel):
         UniqueConstraint("name", "parent_collection_id", name="unique_collection"),
     )
     name: str = Field(index=True)
+    dataset_id: UUID = Field(index=True, foreign_key="dataset.dataset_id")
     parent_collection_id: Optional[UUID] = Field(
-        default=None, foreign_key="collection.collection_id"
+        default=None,
+        foreign_key="collection.collection_id",
+        index=True,
     )
     sample_type: SampleType
 
@@ -38,8 +41,16 @@ class CollectionBase(SQLModel):
     group_component_index: Optional[int] = None
 
 
-class CollectionCreate(CollectionBase):
+class CollectionCreate(SQLModel):
     """Collection class when inserting."""
+
+    name: str
+    parent_collection_id: Optional[UUID] = None
+    sample_type: SampleType
+
+    # Group-specific fields
+    group_component_name: Optional[str] = None
+    group_component_index: Optional[int] = None
 
 
 class CollectionView(CollectionBase):
@@ -62,6 +73,7 @@ class ComponentCollectionView(CollectionBase):
         """Create a ComponentCollectionView from a CollectionTable."""
         return cls(
             name=collection.name,
+            dataset_id=collection.dataset_id,
             parent_collection_id=collection.parent_collection_id,
             sample_type=collection.sample_type,
             group_component_name=collection.group_component_name or "",
@@ -101,3 +113,4 @@ class CollectionTable(CollectionBase, table=True):
         back_populates="parent",
         sa_relationship_kwargs={"lazy": "select"},
     )
+    # TODO(lukas, 3/2026): add a relationship to DatasetTable