+++ lightly_studio/src/lightly_studio/resolvers/collection_resolver/create.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from sqlmodel import Session
 
 from lightly_studio.models.collection import CollectionCreate, CollectionTable
+from lightly_studio.models.dataset import DatasetTable
 from lightly_studio.resolvers import collection_resolver
 
 
@@ -29,8 +30,38 @@ def create(session: Session, collection: CollectionCreate) -> CollectionTable:
     )
     if existing:
         raise ValueError(f"Collection with name '{collection.name}' already exists.")
-    db_collection = CollectionTable.model_validate(collection)
+
+    db_dataset: DatasetTable | None = None
+    if collection.parent_collection_id is None:
+        # If this is the root collection, create a dataset first.
+        db_dataset = DatasetTable()
+        session.add(db_dataset)
+        session.flush([db_dataset])
+        dataset_id = db_dataset.dataset_id
+    else:
+        # Inherit dataset_id from parent collection.
+        parent = collection_resolver.get_by_id(
+            session=session, collection_id=collection.parent_collection_id
+        )
+        if parent is None:
+            raise ValueError(f"Parent collection {collection.parent_collection_id} not found")
+        # Parent exists, get_by_name checks that
+        dataset_id = parent.dataset_id
+
+    # Create the table record, adding the determined dataset_id.
+    db_collection = CollectionTable(
+        **collection.model_dump(),
+        dataset_id=dataset_id,
+    )
+
+    if db_dataset is not None:
+        # Link the dataset back to its root collection. Note that this isn't a foreign key
+        # relationship. Only the collection has a foreign key to the dataset.
+        db_dataset.root_collection_id = db_collection.collection_id
+        session.add(db_dataset)
+
     session.add(db_collection)
     session.commit()
     session.refresh(db_collection)
+
     return db_collection